Make changes to supplied settings visible via pointer - .cpp
User can now provide a pointer to their own SensorSettings structure. If it is supplied it is first filled out to be identical to the desired settings, and then when default cases are selected (because other cases do not match the switch value) the appropriate field is updated in the output SensorSettings. Then the user can double check that SensorSettings matches the settings that they intended. If not then that sub-setting was not a valid choice.