Address issue #5
This completes the handling of issue #5, started with commits 32aca21 and 159af65.

Instead of returning the actual values through the pointer I changed to returning the values that the user wanted through the pointer. Then the actual .settings member of the object is modified to reflect the current settings.

There is now an example to illustrate how to check if the settings have been chnged by the 'begin' method.